cleanup unit tests
