+++ slatedb/src/test_utils.rs
@@ -21,7 +21,7 @@ use object_store::{
 };
 use rand::{Rng, RngCore};
 use std::cmp::Ordering as CmpOrdering;
-use std::collections::{BTreeMap, VecDeque};
+use std::collections::{BTreeMap, HashMap, VecDeque};
 use std::fmt;
 use std::ops::Bound::{Excluded, Included, Unbounded};
 use std::ops::{Bound, RangeBounds};
@@ -33,7 +33,10 @@ use std::thread;
 use std::time::Duration;
 use tokio::sync::Notify;
 use tracing_subscriber::fmt::format::FmtSpan;
+use tracing_subscriber::layer::Context;
+use tracing_subscriber::registry::LookupSpan;
 use tracing_subscriber::EnvFilter;
+use tracing_subscriber::Layer;
 use ulid::Ulid;
 
 pub(crate) fn bounded_sst_view(id: u64, first: &'static [u8], last: &'static [u8]) -> SsTableView {
@@ -80,6 +83,127 @@ pub(crate) async fn assert_next<T: RowEntryIterator>(iterator: &mut T, expected_
 pub(crate) fn assert_kv(kv: &KeyValue, key: &[u8], val: &[u8]) {
     assert_eq!(kv.key, key);
     assert_eq!(kv.value, val);
+}
+
+#[derive(Clone, Debug)]
+pub(crate) struct RecordedSpan {
+    id: tracing::Id,
+    pub(crate) name: String,
+    pub(crate) level: String,
+    pub(crate) parent_name: Option<String>,
+    pub(crate) fields: HashMap<String, String>,
+}
+
+#[derive(Clone, Default)]
+pub(crate) struct SpanRecorder {
+    span_name: Option<&'static str>,
+    spans: Arc<std::sync::Mutex<Vec<RecordedSpan>>>,
+}
+
+impl SpanRecorder {
+    pub(crate) fn for_name(span_name: &'static str) -> Self {
+        Self {
+            span_name: Some(span_name),
+            spans: Arc::new(std::sync::Mutex::new(Vec::new())),
+        }
+    }
+
+    pub(crate) fn spans(&self) -> Vec<RecordedSpan> {
+        self.spans
+            .lock()
+            .expect("The span recorder lock is poisoned.")
+            .clone()
+    }
+
+    pub(crate) fn only_span(&self) -> RecordedSpan {
+        let spans = self.spans();
+        assert_eq!(
+            spans.len(),
+            1,
+            "The recorder must contain one matching span."
+        );
+        spans[0].clone()
+    }
+}
+
+struct SpanFieldRecorder<'a> {
+    fields: &'a mut HashMap<String, String>,
+}
+
+impl tracing::field::Visit for SpanFieldRecorder<'_> {
+    fn record_bool(&mut self, field: &tracing::field::Field, value: bool) {
+        self.fields
+            .insert(field.name().to_string(), value.to_string());
+    }
+
+    fn record_str(&mut self, field: &tracing::field::Field, value: &str) {
+        self.fields
+            .insert(field.name().to_string(), value.to_string());
+    }
+
+    fn record_debug(&mut self, field: &tracing::field::Field, value: &dyn fmt::Debug) {
+        self.fields
+            .insert(field.name().to_string(), format!("{value:?}"));
+    }
+}
+
+impl<S> Layer<S> for SpanRecorder
+where
+    S: tracing::Subscriber + for<'lookup> LookupSpan<'lookup>,
+{
+    fn on_new_span(
+        &self,
+        attrs: &tracing::span::Attributes<'_>,
+        id: &tracing::Id,
+        ctx: Context<'_, S>,
+    ) {
+        if self
+            .span_name
+            .is_some_and(|span_name| attrs.metadata().name() != span_name)
+        {
+            return;
+        }
+        let mut fields = HashMap::new();
+        attrs.record(&mut SpanFieldRecorder {
+            fields: &mut fields,
+        });
+
+        let explicit_parent = attrs.parent().and_then(|id| ctx.span(id));
+        let contextual_parent = ctx.current_span().id().and_then(|id| ctx.span(id));
+        let parent_name = explicit_parent
+            .or(contextual_parent)
+            .map(|span| span.metadata().name().to_string());
+
+        self.spans
+            .lock()
+            .expect("The span recorder lock is poisoned.")
+            .push(RecordedSpan {
+                id: id.clone(),
+                name: attrs.metadata().name().to_string(),
+                level: attrs.metadata().level().to_string(),
+                parent_name,
+                fields,
+            });
+    }
+
+    fn on_record(
+        &self,
+        id: &tracing::Id,
+        values: &tracing::span::Record<'_>,
+        _ctx: Context<'_, S>,
+    ) {
+        let mut fields = HashMap::new();
+        values.record(&mut SpanFieldRecorder {
+            fields: &mut fields,
+        });
+        let mut spans = self
+            .spans
+            .lock()
+            .expect("The span recorder lock is poisoned.");
+        if let Some(span) = spans.iter_mut().find(|span| span.id == *id) {
+            span.fields.extend(fields);
+        }
+    }
 }
 
 pub(crate) struct TestIterator {
